repo-roaster/requirements.txt
openai==1.14.3
GitPython==3.1.42
python-dotenv==1.0.1
rich==13.7.1
click==8.1.7
